+++ runtime/integration-tests/src/integrations_mock/test_common.rs
@@ -983,3 +983,344 @@ fn test_sender_side_xcm_weights() {
     let weight = <XcmExecutorConfig as xcm_executor::Config>::Weigher::weight(&mut msg).unwrap();
     assert!(weight < ADVERTISED_DEST_WEIGHT);
 }
+
+mod governance_tests {
+    use super::{super::mock::ExtBuilder, *};
+    use frame_support::{codec::Encode, traits::StorePreimage};
+    use sp_core::H256;
+    use sp_runtime::traits::{BlakeTwo256, Hash};
+
+    fn note_preimage(proposer: &AccountId, proposal_call: &RuntimeCall) -> H256 {
+        let preimage = proposal_call.encode();
+        let preimage_hash = BlakeTwo256::hash(&preimage[..]);
+        assert_ok!(Preimage::note_preimage(
+            RuntimeOrigin::signed(proposer.clone()),
+            preimage
+        ));
+        preimage_hash
+    }
+
+    fn propose_council_motion(council_motion: &RuntimeCall, proposer: &AccountId) -> H256 {
+        let council_motion_len: u32 = council_motion.using_encoded(|p| p.len() as u32);
+        assert_ok!(Council::propose(
+            RuntimeOrigin::signed(proposer.clone()),
+            1,
+            Box::new(council_motion.clone()),
+            council_motion_len
+        ));
+
+        BlakeTwo256::hash_of(&council_motion)
+    }
+
+    fn start_supermajority_against_governance_assertions(proposer: &AccountId) -> H256 {
+        // Setup the preimage and preimage hash
+        let runtime_call = RuntimeCall::System(frame_system::Call::remark { remark: vec![0] });
+        let preimage_hash = note_preimage(proposer, &runtime_call);
+        let proposal = Preimage::bound(runtime_call).unwrap();
+
+        // Setup the Council and Technical Committee
+        assert_ok!(Council::set_members(
+            root_origin(),
+            vec![proposer.clone()],
+            None,
+            0
+        ));
+        assert_ok!(TechnicalCommittee::set_members(
+            root_origin(),
+            vec![proposer.clone()],
+            None,
+            0
+        ));
+
+        // Setup and propose the Council motion for external_propose_default routine
+        // No voting required because there's only 1 seat.
+        let council_motion =
+            RuntimeCall::Democracy(pallet_democracy::Call::external_propose_default { proposal });
+        let council_motion_hash = propose_council_motion(&council_motion, proposer);
+
+        assert_eq!(
+            last_event(),
+            RuntimeEvent::Council(pallet_collective::Event::Executed {
+                proposal_hash: council_motion_hash,
+                result: Ok(())
+            })
+        );
+
+        preimage_hash
+    }
+
+    fn start_majority_carries_governance_assertions(proposer: &AccountId) -> H256 {
+        let democracy_proposal =
+            RuntimeCall::System(frame_system::Call::remark { remark: vec![0] });
+
+        // Setup the preimage and preimage hash
+        let preimage_hash = note_preimage(
+            proposer,
+            &RuntimeCall::System(frame_system::Call::remark { remark: vec![0] }),
+        );
+
+        // Setup the Council and Technical Committee
+        assert_ok!(Council::set_members(
+            root_origin(),
+            vec![proposer.clone()],
+            None,
+            0
+        ));
+        assert_ok!(TechnicalCommittee::set_members(
+            root_origin(),
+            vec![proposer.clone()],
+            None,
+            0
+        ));
+
+        let bounded_democracy_proposal = Preimage::bound(democracy_proposal).unwrap();
+        // Setup and propose the Council motion for external_propose_default routine
+        // No voting required because there's only 1 seat.
+        let council_motion =
+            RuntimeCall::Democracy(pallet_democracy::Call::external_propose_majority {
+                proposal: bounded_democracy_proposal,
+            });
+        let council_motion_hash = propose_council_motion(&council_motion, proposer);
+
+        assert_eq!(
+            last_event(),
+            RuntimeEvent::Council(pallet_collective::Event::Executed {
+                proposal_hash: council_motion_hash,
+                result: Ok(())
+            })
+        );
+
+        preimage_hash
+    }
+
+    fn end_governance_assertions(
+        referendum_index: u32,
+        end_of_referendum: u32,
+        enactment_period: u32,
+    ) {
+        let time_of_enactment = end_of_referendum + enactment_period;
+        run_to_block(end_of_referendum - 1);
+        assert_eq!(1, Democracy::referendum_count());
+
+        // After the voting period the referendum ends and is scheduled for enactment:
+        run_to_block(end_of_referendum);
+        assert_eq!(
+            last_event(),
+            RuntimeEvent::Scheduler(pallet_scheduler::Event::Scheduled {
+                when: time_of_enactment,
+                index: referendum_index
+            })
+        );
+
+        // After the enactment period the proposal is dispatched:
+        run_to_block(time_of_enactment);
+        assert_eq!(
+            last_event(),
+            RuntimeEvent::Scheduler(pallet_scheduler::Event::Dispatched {
+                task: (time_of_enactment, referendum_index),
+                id: Some([
+                    100, 101, 109, 111, 99, 114, 97, 99, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                    0, 0, 0, 0, 0, 0, 0, 0, 0, 0
+                ]),
+                result: Ok(())
+            })
+        );
+    }
+
+    fn assert_proposal_is_filtered(proposer: &AccountId, motion: &RuntimeCall) {
+        let council_motion_hash = propose_council_motion(motion, proposer);
+
+        assert_eq!(
+            last_event(),
+            RuntimeEvent::Council(pallet_collective::Event::Executed {
+                proposal_hash: council_motion_hash,
+                result: Err(DispatchError::Module(ModuleError {
+                    index: 0,
+                    error: [5, 0, 0, 0],
+                    message: None
+                }))
+            })
+        );
+    }
+
+    #[test]
+    fn fast_track_available() {
+        assert!(<Runtime as pallet_democracy::Config>::InstantAllowed::get());
+    }
+
+    #[test]
+    fn sanity_check_governance_periods() {
+        assert_eq!(LaunchPeriod::get(), 7 * DAYS);
+        assert_eq!(VotingPeriod::get(), 7 * DAYS);
+        assert_eq!(EnactmentPeriod::get(), DAYS);
+    }
+
+    #[test]
+    fn slow_majority_carries_governance_works() {
+        ExtBuilder::default().build().execute_with(|| {
+            let _preimage_hash = start_majority_carries_governance_assertions(&ALICE);
+
+            let start_of_referendum = LaunchPeriod::get();
+            let referendum_index = 0;
+
+            run_to_block(start_of_referendum - 1);
+            assert_eq!(0, Democracy::referendum_count());
+
+            // 7 days in the external proposal queue before the referendum starts.
+            run_to_block(start_of_referendum);
+            assert_eq!(
+                last_event(),
+                RuntimeEvent::Democracy(pallet_democracy::Event::Started {
+                    ref_index: referendum_index,
+                    threshold: pallet_democracy::VoteThreshold::SimpleMajority
+                })
+            );
+            // Time to vote for the referendum with some amount
+            assert_ok!(Democracy::vote(
+                RuntimeOrigin::signed(ALICE.clone()),
+                0,
+                pallet_democracy::AccountVote::Standard {
+                    vote: pallet_democracy::Vote {
+                        aye: true,
+                        conviction: pallet_democracy::Conviction::None
+                    },
+                    balance: 10 * KMA
+                }
+            ));
+
+            end_governance_assertions(
+                referendum_index,
+                start_of_referendum + VotingPeriod::get(),
+                EnactmentPeriod::get(),
+            );
+        });
+    }
+
+    #[test]
+    fn slow_supermajority_against_governance_works() {
+        ExtBuilder::default().build().execute_with(|| {
+            let _preimage_hash = start_supermajority_against_governance_assertions(&ALICE);
+
+            let start_of_referendum = LaunchPeriod::get();
+            let referendum_index = 0;
+
+            run_to_block(start_of_referendum - 1);
+            assert_eq!(0, Democracy::referendum_count());
+
+            // 7 days in the external proposal queue before the referendum starts.
+            run_to_block(start_of_referendum);
+            assert_eq!(1, Democracy::referendum_count());
+            assert_eq!(
+                last_event(),
+                RuntimeEvent::Democracy(pallet_democracy::Event::Started {
+                    ref_index: referendum_index,
+                    threshold: pallet_democracy::VoteThreshold::SuperMajorityAgainst
+                })
+            );
+            // Time to vote for the referendum with some amount
+            assert_ok!(Democracy::vote(
+                RuntimeOrigin::signed(ALICE.clone()),
+                referendum_index,
+                pallet_democracy::AccountVote::Standard {
+                    vote: pallet_democracy::Vote {
+                        aye: true,
+                        conviction: pallet_democracy::Conviction::None
+                    },
+                    balance: 100 * KMA
+                }
+            ));
+
+            end_governance_assertions(
+                referendum_index,
+                start_of_referendum + VotingPeriod::get(),
+                EnactmentPeriod::get(),
+            );
+        });
+    }
+
+    #[test]
+    fn fast_track_governance_works() {
+        ExtBuilder::default().build().execute_with(|| {
+            let preimage_hash = start_supermajority_against_governance_assertions(&ALICE);
+
+            let voting_period = 5;
+            let enactment_period = 5;
+            let referendum_index = 0;
+
+            // Setup and propose the Technical Committee motion for the fast_track routine
+            // No voting required because there's only 1 seat.
+            // Voting and delay periods of 5 blocks so this should be enacted on block 11
+            let tech_committee_motion =
+                RuntimeCall::Democracy(pallet_democracy::Call::fast_track {
+                    proposal_hash: preimage_hash,
+                    voting_period,
+                    delay: enactment_period,
+                });
+            let tech_committee_motion_len: u32 =
+                tech_committee_motion.using_encoded(|p| p.len() as u32);
+            let tech_committee_motion_hash = BlakeTwo256::hash_of(&tech_committee_motion);
+            assert_ok!(TechnicalCommittee::propose(
+                RuntimeOrigin::signed(ALICE.clone()),
+                1,
+                Box::new(tech_committee_motion),
+                tech_committee_motion_len
+            ));
+            // Make sure the motion was actually executed
+            assert_eq!(
+                last_event(),
+                RuntimeEvent::TechnicalCommittee(pallet_collective::Event::Executed {
+                    proposal_hash: tech_committee_motion_hash,
+                    result: Ok(())
+                })
+            );
+
+            // Time to vote for the referendum with some amount
+            assert_ok!(Democracy::vote(
+                RuntimeOrigin::signed(ALICE.clone()),
+                referendum_index,
+                pallet_democracy::AccountVote::Standard {
+                    vote: pallet_democracy::Vote {
+                        aye: true,
+                        conviction: pallet_democracy::Conviction::None
+                    },
+                    balance: 10 * KMA
+                }
+            ));
+
+            // No launch period because of the fast track.
+            end_governance_assertions(
+                referendum_index,
+                System::block_number() + voting_period,
+                enactment_period,
+            );
+        });
+    }
+
+    #[test]
+    fn governance_filters_work() {
+        assert!(<Runtime as pallet_democracy::Config>::InstantAllowed::get());
+
+        ExtBuilder::default().build().execute_with(|| {
+            // Setup the preimage and preimage hash
+            let runtime_call = RuntimeCall::System(frame_system::Call::remark { remark: vec![0] });
+            let proposal = Preimage::bound(runtime_call).unwrap();
+
+            // Setup the Council
+            assert_ok!(Council::set_members(
+                root_origin(),
+                vec![ALICE.clone()],
+                None,
+                0
+            ));
+
+            // Public proposals should be filtered out.
+            assert_proposal_is_filtered(
+                &ALICE,
+                &RuntimeCall::Democracy(pallet_democracy::Call::propose {
+                    proposal,
+                    value: 100 * KMA,
+                }),
+            );
+        });
+    }
+}
